MCP-CrewLink/requirements.txt
# Core MCP dependencies
mcp>=1.0.0
pydantic>=2.0.0

# CrewAI for AI agents and tasks
crewai>=0.28.0
crewai-tools>=0.12.0

# OpenAI API for image generation
openai>=1.12.0

# HTTP client for API requests
httpx>=0.25.0

# Additional utilities
python-dotenv>=1.0.0
wandb>=0.16.0
random2>=1.0.1
psutil>=5.9.0

# Development and debugging (optional)
jupyter>=1.0.0
ipython>=8.0.0

# Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Type checking (optional)
mypy>=1.7.0
types-requests>=2.31.0